// src/types/IslandTiles.ts
export type IslandTiles =
    'center1'
    | 'center2'
    | 'center3'
    | 'center4'
    | 'bottom1'
    | 'bottom2'
    | 'bottomLeft'
    | 'bottomRight'
    | 'left1'
    | 'left2'
    | 'right1'
    | 'right2'
    | 'top1'
    | 'top2'
    | 'topLeft'
    | 'topRight'
    | 'innerTopRight'
    | 'innerTopLeft'
    | 'innerBottomRight'
    | 'innerBottomLeft'